requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Markdown==3.2.1
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
